requirements.txt
streamlit 
PyPDF2 
pdfplumber 
faiss-cpu 
sentence-transformers 
transformers
langchain
langchain-community
openai
python-dotenv
tiktoken
pypdf
pillow
PyMuPDF
streamlit-pdf-viewer